Fix errors on Helfer delete

+++ web/modules/custom/startklar/src/EventSubscriber/ReferenceDeletionSubscriber.php
@@ -58,10 +58,8 @@ class ReferenceDeletionSubscriber implements EventSubscriberInterface {
       }
 
       // Delete person nodes when a helfer anmeldung is deleted
-      if ($entity->hasField('field_person')) {
-        if ($entity->field_person->entity) {
+      if ($entity->hasField('field_person') && !$entity->field_person->isEmpty()) {
           $fieldItem->entity->delete();
-        }
       }
     }
   }
